Fix most remaining LGTM alerts (#1893)
* fixups from LGTM

* short globals

* fix spinner and textwindow

* total ordering

* no spinner/text window when import from manager

* not android
old-commit-hash: 5db81f60ca8fbc0ef28924b7e93f25a1d4bb3764

+++ selfdrive/modeld/models/commonmodel.c
@@ -15,13 +15,13 @@ void frame_init(ModelFrame* frame, int width, int height,
   frame->transformed_height = height;
 
   frame->transformed_y_cl = clCreateBuffer(frame->context, CL_MEM_READ_WRITE,
-                                       frame->transformed_width*frame->transformed_height, NULL, &err);
+                                           (size_t)frame->transformed_width*frame->transformed_height, NULL, &err);
   assert(err == 0);
   frame->transformed_u_cl = clCreateBuffer(frame->context, CL_MEM_READ_WRITE,
-                                       (frame->transformed_width/2)*(frame->transformed_height/2), NULL, &err);
+                                           (size_t)(frame->transformed_width/2)*(frame->transformed_height/2), NULL, &err);
   assert(err == 0);
   frame->transformed_v_cl = clCreateBuffer(frame->context, CL_MEM_READ_WRITE,
-                                       (frame->transformed_width/2)*(frame->transformed_height/2), NULL, &err);
+                                           (size_t)(frame->transformed_width/2)*(frame->transformed_height/2), NULL, &err);
   assert(err == 0);
 
   frame->net_input_size = ((width*height*3)/2)*sizeof(float);
